+++ tripaille/commands/cmd_expression.py
@@ -1,6 +1,6 @@
 import click
 from tripaille.commands.expression.add_expression import cli as func0
-
+from tripaille.commands.expression.add_biomaterial import cli as func1
 
 @click.group()
 def cli():
@@ -9,3 +9,4 @@ def cli():
 
 
 cli.add_command(func0)
+cli.add_command(func1)
