+++ tripal/expression/__init__.py
@@ -87,3 +87,41 @@ class ExpressionClient(Client):
             return r
         else:
             return self._run_job_and_wait(r['job_id'])
+
+    def add_biomaterial(self, organism_id, file_path, file_type, no_wait=False):
+        """
+        Add a new biomaterial to the database
+
+        :type organism_id: str
+        :param organism_id: The id of the associated organism
+
+        :type file_path: str
+        :param file_path: The path to the biomaterial file
+
+        :type file_type: str
+        :param file_type: The type of the biomaterial file (xml, tsv or csv)
+
+        :rtype: dict
+        :return: Job information
+        """
+
+        if file_type not in ['xml', 'tsv', 'csv']:
+            raise Exception("File format must be one of ['xml', 'tsv', 'csv']")
+
+        if file_type == "xml":
+            job_args = [file_path, organism_id]
+            callback = 'xml_biomaterial_parser'
+
+        else:
+            job_args = [file_path, organism_id, file_type]
+            callback = 'flat_biomaterial_parser'
+
+        r = self.tripal.job.add_job("Add Biomaterial", 'tripal_analysis_expression', callback, job_args)
+
+        if 'job_id' not in r or not r['job_id']:
+            raise Exception("Failed to create job, received %s" % r)
+
+        if no_wait:
+            return r
+        else:
+            return self._run_job_and_wait(r['job_id'])
